server/requirements.txt
aiohttp~=3.6.2
PyYAML~=5.3.1
multipart~=0.2.1
elasticsearch-dsl>=7.0.0,<8.0.0
elasticsearch~=7.8.1
certifi~=2020.6.20
chardet~=3.0.4
netaddr~=0.8.0
formatflowed~=2.0.0
requests~=2.24.0
google-auth~=1.21.1
